webapp/requirements.txt
Django>=1.7,<1.8
psycopg2>=2.6
pytz
djangorestframework
markdown
django-filter
pinax-notifications>=2.0.1
